+++ test/Fixture/Classes/FinalRuleWithExcludedClassNames/Success/FinalClassWithAnonymousClass.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Localheinz\PHPStan\Rules\Test\Fixture\Classes\AbstractOrFinalRule\Success;
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Classes\FinalRuleWithExcludedClassNames\Success;
 
-trait TraitWithAnonymousClass
+final class FinalClassWithAnonymousClass
 {
     public function create()
     {